Don't delete repos or reference folders

+++ src/aoietl/copy_to_fileshare.py
@@ -19,7 +19,7 @@ def clear_azure_file_share(share_client):
             for item in items:
                 item_path = f"{dir_path}/{item['name']}" if dir_path else item['name']
                 
-                if item['is_directory']:
+                if item['is_directory'] and item['name'] not in ["reference", "repos"]:
                     # It's a directory - recursively delete its contents first
                     sub_dir_client = share_client.get_directory_client(item_path)
                     delete_directory_contents(sub_dir_client, item_path)
